log: move nvme_{set|get}_debug to log.c
Let's group all logging related function into one location.
